+++ packages/core/src/browser/menu/browser-main-menu-factory.ts
@@ -17,18 +17,16 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 import { injectable, inject } from 'inversify';
-import { MenuBar, Menu as MenuWidget, Widget } from '@phosphor/widgets';
+import { MenuBar, Menu as MenuWidget } from '@phosphor/widgets';
 import { CommandRegistry as PhosphorCommandRegistry } from '@phosphor/commands';
 import {
     CommandRegistry, ActionMenuNode, CompositeMenuNode,
-    MenuModelRegistry, MAIN_MENU_BAR, MenuPath, ILogger
+    MenuModelRegistry, MAIN_MENU_BAR, MenuPath
 } from '../../common';
 import { KeybindingRegistry } from '../keybinding';
-import { FrontendApplicationContribution, FrontendApplication } from '../frontend-application';
 import { ContextKeyService } from '../context-key-service';
 import { ContextMenuContext } from './context-menu-context';
 import { waitForRevealed } from '../widgets';
-import { ApplicationShell } from '../shell';
 
 export abstract class MenuBarWidget extends MenuBar {
     abstract activateMenu(label: string, ...labels: string[]): Promise<MenuWidget>;
@@ -38,20 +36,20 @@ export abstract class MenuBarWidget extends MenuBar {
 @injectable()
 export class BrowserMainMenuFactory {
 
-    @inject(ILogger)
-    protected readonly logger: ILogger;
-
     @inject(ContextKeyService)
     protected readonly contextKeyService: ContextKeyService;
 
     @inject(ContextMenuContext)
     protected readonly context: ContextMenuContext;
 
-    constructor(
-        @inject(CommandRegistry) protected readonly commandRegistry: CommandRegistry,
-        @inject(KeybindingRegistry) protected readonly keybindingRegistry: KeybindingRegistry,
-        @inject(MenuModelRegistry) protected readonly menuProvider: MenuModelRegistry
-    ) { }
+    @inject(CommandRegistry)
+    protected readonly commandRegistry: CommandRegistry;
+
+    @inject(KeybindingRegistry)
+    protected readonly keybindingRegistry: KeybindingRegistry;
+
+    @inject(MenuModelRegistry)
+    protected readonly menuRegistry: MenuModelRegistry;
 
     createMenuBar(): MenuBarWidget {
         const menuBar = new DynamicMenuBarWidget();
@@ -65,8 +63,8 @@ export class BrowserMainMenuFactory {
         return menuBar;
     }
 
-    protected fillMenuBar(menuBar: MenuBarWidget): void {
-        const menuModel = this.menuProvider.getMenu(MAIN_MENU_BAR);
+    fillMenuBar(menuBar: MenuBarWidget): void {
+        const menuModel = this.menuRegistry.getMenu(MAIN_MENU_BAR);
         const phosphorCommands = this.createPhosphorCommands(menuModel);
         // for the main menu we want all items to be visible.
         phosphorCommands.isVisible = () => true;
@@ -80,7 +78,7 @@ export class BrowserMainMenuFactory {
     }
 
     createContextMenu(path: MenuPath, args?: any[]): MenuWidget {
-        const menuModel = this.menuProvider.getMenu(path);
+        const menuModel = this.menuRegistry.getMenu(path);
         const phosphorCommands = this.createPhosphorCommands(menuModel, args);
 
         const contextMenu = new DynamicMenuWidget(menuModel, { commands: phosphorCommands }, this.contextKeyService, this.context);
@@ -297,33 +295,4 @@ class DynamicMenuWidget extends MenuWidget {
         return items;
     }
 
-}
-
-@injectable()
-export class BrowserMenuBarContribution implements FrontendApplicationContribution {
-
-    @inject(ApplicationShell)
-    protected readonly shell: ApplicationShell;
-
-    constructor(
-        @inject(BrowserMainMenuFactory) protected readonly factory: BrowserMainMenuFactory
-    ) { }
-
-    onStart(app: FrontendApplication): void {
-        const logo = this.createLogo();
-        app.shell.addWidget(logo, { area: 'top' });
-        const menu = this.factory.createMenuBar();
-        app.shell.addWidget(menu, { area: 'top' });
-    }
-
-    get menuBar(): MenuBarWidget | undefined {
-        return this.shell.topPanel.widgets.find(w => w instanceof MenuBarWidget) as MenuBarWidget | undefined;
-    }
-
-    protected createLogo(): Widget {
-        const logo = new Widget();
-        logo.id = 'theia:icon';
-        logo.addClass('theia-icon');
-        return logo;
-    }
 }